Make TokamakDOM depend on TokamakStaticHTML
