+++ Sources/TokamakDOM/Styles/ToggleStyle.swift
@@ -16,6 +16,7 @@
 //
 
 import TokamakCore
+import TokamakStaticHTML
 
 public struct DefaultToggleStyle: ToggleStyle {
   public func makeBody(configuration: Configuration) -> some View {
@@ -32,7 +33,7 @@ public struct CheckboxToggleStyle: ToggleStyle {
       attrs["checked"] = "checked"
     }
     return HTML("label") {
-      HTML("input", attrs, listeners: [
+      DynamicHTML("input", attrs, listeners: [
         "change": { event in
           let checked = event.target.object?.checked.boolean ?? false
           configuration.isOn = checked
