+++ Sources/TokamakStaticDemo/main.swift
@@ -1,17 +1,21 @@
+// Copyright 2020 Tokamak contributors
 //
-//  File.swift
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 //
 //  Created by Carson Katri on 7/20/20.
 //
 
-import TokamakStatic
-
-struct ContentView: View {
-  var body: some View {
-    Text("Hello, world!")
-  }
-}
+import TokamakStaticHTML
 
 let renderer = StaticRenderer(ContentView())
 print(renderer.html)
