+++ Sources/TokamakStaticHTML/Resources/TokamakStyles.swift
@@ -55,7 +55,7 @@ public let tokamakStyles = """
 }
 """
 
-let rootNodeStyles = """
+public let rootNodeStyles = """
 display: flex;
 width: 100%;
 height: 100%;
