+++ Sources/TokamakStaticHTML/Shapes/Path.swift
