+++ Sources/TokamakStaticHTML/Shapes/Shape.swift
