+++ Sources/TokamakStaticHTML/Shapes/_ShapeView.swift
