+++ Sources/TokamakStaticHTML/Views/Layout/LazyVGrid.swift
@@ -17,11 +17,11 @@
 
 import TokamakCore
 
-extension LazyHGrid: SpacerContainer {
-  var axis: SpacerContainerAxis { .horizontal }
-  var hasSpacer: Bool { false }
-  var fillCrossAxis: Bool {
-    _LazyHGridProxy(self).rows.contains {
+extension LazyVGrid: SpacerContainer {
+  public var axis: SpacerContainerAxis { .vertical }
+  public var hasSpacer: Bool { false }
+  public var fillCrossAxis: Bool {
+    _LazyVGridProxy(self).columns.contains {
       if case .adaptive(minimum: _, maximum: _) = $0.size {
         return true
       } else {
@@ -31,31 +31,31 @@ extension LazyHGrid: SpacerContainer {
   }
 }
 
-extension LazyHGrid: ViewDeferredToRenderer {
-  var lastRow: GridItem? {
-    _LazyHGridProxy(self).rows.last
+extension LazyVGrid: ViewDeferredToRenderer {
+  public var lastColumn: GridItem? {
+    _LazyVGridProxy(self).columns.last
   }
 
   public var deferredBody: AnyView {
     var styles = """
     display: grid;
-    grid-template-rows: \(_LazyHGridProxy(self)
-      .rows
+    grid-template-columns: \(_LazyVGridProxy(self)
+      .columns
       .map(\.description)
       .joined(separator: " "));
-    grid-auto-flow: column;
+    grid-auto-flow: row;
     """
     if fillCrossAxis {
-      styles += "height: 100%;"
+      styles += "width: 100%;"
     }
-    // CSS Grid doesn't let these be specified for specific rows
-    if let lastRow = lastRow {
-      styles += "justify-items: \(lastRow.alignment.horizontal.cssValue);"
-      styles += "align-items: \(lastRow.alignment.vertical.cssValue);"
+    // CSS Grid doesn't let these be specified for specific columns
+    if let lastCol = lastColumn {
+      styles += "justify-items: \(lastCol.alignment.horizontal.cssValue);"
+      styles += "align-items: \(lastCol.alignment.vertical.cssValue);"
     }
-    styles += "grid-gap: \(_LazyHGridProxy(self).spacing)px;"
+    styles += "grid-gap: \(_LazyVGridProxy(self).spacing)px;"
     return AnyView(HTML("div", ["style": styles]) {
-      _LazyHGridProxy(self).content
+      _LazyVGridProxy(self).content
     })
   }
 }
