+++ Sources/TokamakStaticHTML/Views/Spacers/Spacer.swift
@@ -14,18 +14,18 @@
 
 import TokamakCore
 
-enum SpacerContainerAxis {
+public enum SpacerContainerAxis {
   case horizontal, vertical
 }
 
-protocol SpacerContainer {
+public protocol SpacerContainer {
   var hasSpacer: Bool { get }
   var axis: SpacerContainerAxis { get }
   var fillCrossAxis: Bool { get }
 }
 
 extension SpacerContainer where Self: ParentView {
-  var hasSpacer: Bool {
+  public var hasSpacer: Bool {
     children
       .compactMap {
         mapAnyView($0) { (v: Spacer) in
@@ -45,7 +45,7 @@ extension SpacerContainer where Self: ParentView {
   // Does a child SpacerContainer along the opposite axis have a spacer?
   // (e.g., an HStack with a child VStack which contains a spacer)
   // If so, we need to fill the cross-axis so the child can show the correct layout.
-  var fillCrossAxis: Bool {
+  public var fillCrossAxis: Bool {
     children
       .compactMap {
         mapAnyView($0) { (v: SpacerContainer) in v }
